Refactor incorrectly snake cased file name
